refactor: divide time capsule detail screen into components
